requirements.txt
asgiref==3.8.1
Django==4.2.16
sqlparse==0.5.1
typing_extensions==4.12.2
psycopg2-binary==2.9.9
Pillow==10.4.0
django-modeltranslation==0.19.9
django-ckeditor==6.7.1
djangorestframework==3.15.2
drf-spectacular==0.27.2
python-decouple==3.8
